+++ crates/cli/tests/alias_test.rs
@@ -1,7 +1,8 @@
 mod utils;
 
 use predicates::prelude::*;
-use std::fs;
+use proto_core::Manifest;
+use rustc_hash::FxHashMap;
 use utils::*;
 
 #[test]
@@ -20,15 +21,12 @@ fn updates_manifest_file() {
         .success();
 
     assert!(manifest_file.exists());
+
+    let manifest = Manifest::load(manifest_file).unwrap();
+
     assert_eq!(
-        fs::read_to_string(manifest_file).unwrap(),
-        r#"{
-  "aliases": {
-    "example": "19.0.0"
-  },
-  "default_version": null,
-  "installed_versions": []
-}"#
+        manifest.aliases,
+        FxHashMap::from_iter([("example".into(), "19.0.0".into())])
     );
 }
 
@@ -37,18 +35,9 @@ fn can_overwrite_existing_alias() {
     let temp = create_temp_dir();
     let manifest_file = temp.join("tools/node/manifest.json");
 
-    fs::create_dir_all(manifest_file.parent().unwrap()).unwrap();
-    fs::write(
-        &manifest_file,
-        r#"{
-  "aliases": {
-    "example": "19.0.0"
-  },
-  "default_version": null,
-  "installed_versions": []
-}"#,
-    )
-    .unwrap();
+    let mut manifest = Manifest::load(&manifest_file).unwrap();
+    manifest.aliases.insert("example".into(), "19.0.0".into());
+    manifest.save().unwrap();
 
     let mut cmd = create_proto_command(temp.path());
     cmd.arg("alias")
@@ -58,15 +47,11 @@ fn can_overwrite_existing_alias() {
         .assert()
         .success();
 
+    let manifest = Manifest::load(&manifest_file).unwrap();
+
     assert_eq!(
-        fs::read_to_string(manifest_file).unwrap(),
-        r#"{
-  "aliases": {
-    "example": "20.0.0"
-  },
-  "default_version": null,
-  "installed_versions": []
-}"#
+        manifest.aliases,
+        FxHashMap::from_iter([("example".into(), "20.0.0".into())])
     );
 }