+++ crates/cli/tests/install_lang_test.rs
@@ -1,6 +1,8 @@
 mod utils;
 
 use predicates::prelude::*;
+use proto_core::Manifest;
+use rustc_hash::FxHashSet;
 use std::fs;
 use utils::*;
 
@@ -65,15 +67,12 @@ mod node {
         assert!(temp.join("tools/node/19.0.0").exists());
         assert!(temp.join("tools/npm/8.19.2").exists());
 
+        let manifest = Manifest::load(temp.join("tools/npm/manifest.json")).unwrap();
+
+        assert_eq!(manifest.default_version, Some("bundled".into()));
         assert_eq!(
-            fs::read_to_string(temp.join("tools/npm/manifest.json")).unwrap(),
-            r#"{
-  "aliases": {},
-  "default_version": "bundled",
-  "installed_versions": [
-    "8.19.2"
-  ]
-}"#
+            manifest.installed_versions,
+            FxHashSet::from_iter(["8.19.2".into()])
         );
     }